delete is working...finally
