+++ server/models/Job.model.js
@@ -2,7 +2,6 @@ const mongoose = require('mongoose');
 
 
 let jobSchema = mongoose.Schema({
-    id: Number,
     company: String,
     link: String,
     notes: String,
